requirements.txt
flask==3.0.3
flask-bcrypt==1.0.1
PyJWT==2.8.0
pymongo==4.7.0
boto3==1.34.84
pandas==2.2.1
reportlab==4.1.0
cryptography==42.0.5
